requirements.txt
# requirements.txt

# Core Application Frameworks
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.45.1

# Machine Learning & Data Handling
scikit-learn==1.5.2
joblib==1.5.0
pandas==2.2.2
numpy==1.26.4
nltk==3.9.1

# API Interaction & Environment
openai==1.78.1
python-dotenv==1.0.1
requests==2.32.3

# Pydantic & Related
pydantic==2.11.4
starlette==0.38.6

# HTTP & Networking
httpx>=0.24.1
certifi>=2021.10.8
charset-normalizer>=2.0.0
idna>=2.8
urllib3>=1.26.0
sniffio>=1.1.0
tenacity>=8.0.1

# Streamlit Dependencies
altair>=5.0.0
blinker>=1.6.2
cachetools>=4.0
click>=7.0
Jinja2>=3.0
MarkupSafe>=2.0
markdown-it-py>=2.2.0
Pygments>=2.11.0
pillow>=9.0.0
protobuf>=3.20
pyarrow>=10.0.0
python-dateutil>=2.8.0
pytz
rich>=10.0.0
toml>=0.10.0
tornado>=6.1
typer[all]==0.15.3

# NLTK
regex>=2022.1.18

# Scikit-learn/Numpy
scipy>=1.7.3
threadpoolctl>=2.0.0

# Watchfiles for Uvicorn reload
watchfiles>=0.13